+++ backend-app/TestControllers/DeviceTestController.cs
@@ -324,7 +324,10 @@ namespace backEndApp.TestControllers {
             if(!_sensorRepository.DeleteSensors(sensorsToDelete.ToList())) {
                 ModelState.AddModelError("", "Something went wrong when deleting Sensors");
             }
-
+            if(deviceToDelete == null) {
+                ModelState.AddModelError("", "Something went wrong when getting the Device");
+                return StatusCode(500, ModelState);
+            }
             if(!_deviceRepository.DeleteDevice(deviceToDelete)) {
                 ModelState.AddModelError("", "Something went wrong when deleting the Device");
             }
